hugo: cleanup and migrastions
